Raise cython activation values to double precision (64-bit) An attempt to avoid nondeterministic behaviour on the HEC

+++ utils/maths_core.pyx
@@ -20,12 +20,14 @@ caiwingfield.net
 import cython
 from libc.math cimport exp, sqrt, pi, log, erf
 
-TAU: cython.float = 2 * pi
+# cython.double matches the precision of python's float
+
+TAU: cython.double = 2 * pi
 
 
 def exponential_decay(age: cython.uint,
-                      original_activation: cython.float,
-                      decay_factor: cython.float) -> cython.float:
+                      original_activation: cython.double,
+                      decay_factor: cython.double) -> cython.double:
     """
     :param age:
     :param original_activation:
@@ -36,12 +38,12 @@ def exponential_decay(age: cython.uint,
 
 
 def gaussian_decay(age: cython.uint,
-                   original_activation: cython.float,
-                   height_coef: cython.float,
-                   reset_height: cython.float,
-                   centre: cython.float,
-                   sd: cython.float,
-                   ) -> cython.float:
+                   original_activation: cython.double,
+                   height_coef: cython.double,
+                   reset_height: cython.double,
+                   centre: cython.double,
+                   sd: cython.double,
+                   ) -> cython.double:
     """
     :param age:
     :param original_activation:
@@ -55,7 +57,7 @@ def gaussian_decay(age: cython.uint,
     return original_activation * height_coef * reset_height * gaussian_pdf(x=age, mu=centre, sd=sd)
 
 
-def gaussian_pdf(x: cython.float, mu: cython.float, sd: cython.float) -> cython.float:
+def gaussian_pdf(x: cython.double, mu: cython.double, sd: cython.double) -> cython.double:
     """
     Cythonised version of scipy.stats.norm.pdf for scalars.
     :param x:
@@ -63,16 +65,16 @@ def gaussian_pdf(x: cython.float, mu: cython.float, sd: cython.float) -> cython.
     :param sd:
     :return:
     """
-    y:        cython.float = (x - mu) / sd
-    exponent: cython.float = (-1) * y * y / 2
-    denom:    cython.float = sqrt(TAU)
+    y:        cython.double = (x - mu) / sd
+    exponent: cython.double = (-1) * y * y / 2
+    denom:    cython.double = sqrt(TAU)
 
-    e_term:   cython.float = exp(exponent) / denom
+    e_term:   cython.double = exp(exponent) / denom
 
     return e_term / sd
 
 
-def lognormal_sf(x: cython.float, mu: cython.float = 0, sigma: cython.float = 1) -> cython.float:
+def lognormal_sf(x: cython.double, mu: cython.double = 0, sigma: cython.double = 1) -> cython.double:
     """
     Cythonised approximation of the lognormal sf.
     Assumes mu is 0.
@@ -86,7 +88,7 @@ def lognormal_sf(x: cython.float, mu: cython.float = 0, sigma: cython.float = 1)
     return 1 - lognormal_cdf(x, mu, sigma)
 
 
-def lognormal_cdf(x: cython.float, mu: cython.float = 0, sigma: cython.float = 1) -> cython.float:
+def lognormal_cdf(x: cython.double, mu: cython.double = 0, sigma: cython.double = 1) -> cython.double:
     """
     Cythonised approximation of the lognormal cdf.
     :param x:
@@ -96,13 +98,13 @@ def lognormal_cdf(x: cython.float, mu: cython.float = 0, sigma: cython.float = 1
         SD of the log of the random variable.
     :return:
     """
-    numerator:   cython.float = log(x) - mu
-    denomenator: cython.float = sqrt(2) * sigma
-    fraction:    cython.float = numerator / denomenator
+    numerator:   cython.double = log(x) - mu
+    denomenator: cython.double = sqrt(2) * sigma
+    fraction:    cython.double = numerator / denomenator
     return 0.5 + 0.5 * erf(fraction)
 
 
-def lognormal_pdf(x: cython.float, mu: cython.float = 0, sigma: cython.float = 1) -> cython.float:
+def lognormal_pdf(x: cython.double, mu: cython.double = 0, sigma: cython.double = 1) -> cython.double:
     """
     Cythonised approximation of the lognormal pdf.
     :param x:
@@ -110,9 +112,9 @@ def lognormal_pdf(x: cython.float, mu: cython.float = 0, sigma: cython.float = 1
     :param sigma:
     :return:
     """
-    coefficient: cython.float = 1 / (sigma * x * sqrt(TAU))
-    log_term:    cython.float = log(x) - mu
-    numerator:   cython.float = log_term * log_term
-    denomenator: cython.float = 2 * sigma * sigma
-    exponent:    cython.float = (-1) * numerator / denomenator
+    coefficient: cython.double = 1 / (sigma * x * sqrt(TAU))
+    log_term:    cython.double = log(x) - mu
+    numerator:   cython.double = log_term * log_term
+    denomenator: cython.double = 2 * sigma * sigma
+    exponent:    cython.double = (-1) * numerator / denomenator
     return coefficient * exp(exponent)
